x/luqchain/types/message_submit_val_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"luqchain/testutil/sample"
)

func TestMsgSubmitVal_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgSubmitVal
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgSubmitVal{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgSubmitVal{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
